Rename containers.rs to js_collections.rs to match prior implementation
