+++ src/objects/impls/construction_site.rs
@@ -1,6 +1,6 @@
 use crate::{
     constants::{ReturnCode, StructureType},
-    containers::JsContainerFromValue,
+    js_collections::JsContainerFromValue,
     objects::{Owner, RoomObject},
     prelude::*,
 };
